working selector and tooltip
push before working on radar chart
